consolidate gendocs into a single docs-generating package
