+++ cmd/docs-generating/gendocs/gen_kubectl_docs.go
@@ -20,12 +20,12 @@ import (
 	"bytes"
 	"fmt"
 	"io"
+	"k8s.io/kubernetes/cmd/docs-generating/genutils"
 	"os"
 
 	"github.com/spf13/cobra/doc"
 	"k8s.io/cli-runtime/pkg/genericiooptions"
 	"k8s.io/kubectl/pkg/cmd"
-	"k8s.io/kubernetes/cmd/genutils"
 )
 
 func main() {
