+++ hack/update-generated-docs.sh
@@ -28,10 +28,10 @@ source "${KUBE_ROOT}/hack/lib/init.sh"
 kube::golang::setup_env
 
 BINS=(
-	cmd/gendocs
-	cmd/genkubedocs
-	cmd/genman
-	cmd/genyaml
+	cmd/docs-generating/gendocs
+	cmd/docs-generating/genkubedocs
+	cmd/docs-generating/genman
+	cmd/docs-generating/genyaml
 )
 make -C "${KUBE_ROOT}" WHAT="${BINS[*]}"
